+++ src/js/index.js
@@ -1,3 +1,14 @@
+drivershub = `    ____       _                         __  __      __  
+   / __ \\_____(_)   _____  __________   / / / /_  __/ /_ 
+  / / / / ___/ / | / / _ \\/ ___/ ___/  / /_/ / / / / __ \\
+ / /_/ / /  / /| |/ /  __/ /  (__  )  / __  / /_/ / /_/ /
+/_____/_/  /_/ |___/\\___/_/  /____/  /_/ /_/\\__,_/_.___/ 
+                                                         `
+console.log(drivershub);
+console.log("Drivers Hub: Frontend (v1.3.5)");
+console.log("Copyright (C) 2022 CharlesWithC All rights reserved.");
+console.log('This product must work with "Drivers Hub: Backend" which is also made by CharlesWithC!');
+
 VERIFIED = `<svg style="display:inline;position:relative;top:-1.5px;color:skyblue" width="18" height="18" stroke-width="1.5" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg"> <path d="M10.5213 2.62368C11.3147 1.75255 12.6853 1.75255 13.4787 2.62368L14.4989 3.74391C14.8998 4.18418 15.4761 4.42288 16.071 4.39508L17.5845 4.32435C18.7614 4.26934 19.7307 5.23857 19.6757 6.41554L19.6049 7.92905C19.5771 8.52388 19.8158 9.10016 20.2561 9.50111L21.3763 10.5213C22.2475 11.3147 22.2475 12.6853 21.3763 13.4787L20.2561 14.4989C19.8158 14.8998 19.5771 15.4761 19.6049 16.071L19.6757 17.5845C19.7307 18.7614 18.7614 19.7307 17.5845 19.6757L16.071 19.6049C15.4761 19.5771 14.8998 19.8158 14.4989 20.2561L13.4787 21.3763C12.6853 22.2475 11.3147 22.2475 10.5213 21.3763L9.50111 20.2561C9.10016 19.8158 8.52388 19.5771 7.92905 19.6049L6.41553 19.6757C5.23857 19.7307 4.26934 18.7614 4.32435 17.5845L4.39508 16.071C4.42288 15.4761 4.18418 14.8998 3.74391 14.4989L2.62368 13.4787C1.75255 12.6853 1.75255 11.3147 2.62368 10.5213L3.74391 9.50111C4.18418 9.10016 4.42288 8.52388 4.39508 7.92905L4.32435 6.41553C4.26934 5.23857 5.23857 4.26934 6.41554 4.32435L7.92905 4.39508C8.52388 4.42288 9.10016 4.18418 9.50111 3.74391L10.5213 2.62368Z" stroke="currentColor" stroke-width="1.5"/> <path d="M9 12L11 14L15 10" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"/> </svg> `;
 LOCKED = `<svg style="display:inline;position:relative;top:-1.5px;color:red" xmlns="http://www.w3.org/2000/svg" width="18" height="18" enable-background="new 0 0 24 24" viewBox="0 0 24 24"><path d="M17,9V7c0-2.8-2.2-5-5-5S7,4.2,7,7v2c-1.7,0-3,1.3-3,3v7c0,1.7,1.3,3,3,3h10c1.7,0,3-1.3,3-3v-7C20,10.3,18.7,9,17,9z M9,7c0-1.7,1.3-3,3-3s3,1.3,3,3v2H9V7z M13,17c0,0.6-0.4,1-1,1s-1-0.4-1-1v-3c0-0.6,0.4-1,1-1s1,0.4,1,1V17z" fill="red"/></svg>`;
 
@@ -422,7 +433,10 @@ function validate() {
               </svg>&nbsp;&nbsp;<span id="livedriver2" style="color:${color}"></span><span style="color:orange"></p>`);
             }
         }, error: function(data){
-            window.location.href = "/login";
+            if(parseInt(data.status) == 401){
+                localStorage.removeItem("token");
+                window.location.href = "/login";
+            }
         }
     });
     $.ajax({
@@ -639,6 +653,7 @@ function loadDistanceUnit() {
 
 $(document).ready(function () {
     loadDistanceUnit();
+    PreserveApplicationQuestion();
     if (localStorage.getItem("darkmode") == "1") {
         $("body").addClass("bg-gray-800");
         $("body").css("color", "white");
@@ -746,6 +761,7 @@ $(document).ready(function () {
     });
 
     function devwarn() {
+        return;
         for (var i = 0; i < 3; i++) {
             setTimeout(function () {
                 console.log("%cHold Up!", "color: #0000ff; font-size: 100px;");